requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
pyyaml>=6.0
requests>=2.31.0
pillow>=10.0.0
transformers>=4.50.0
torch>=2.0.0
marker-pdf>=0.2.0
python-dotenv>=1.0.0